Refactor: count와 times 자료형 변경(int->long)
- n의 범위가 명시되어있지 않음으로, long으로 자료형을 변경하여 예외에 대처

+++ src/main/java/racingcar/domain/Race.java
@@ -5,13 +5,13 @@ import racingcar.utils.RandomNumberGenerater;
 import java.util.List;
 
 public class Race {
-    int count;
-    int times;
+    long count;
+    long times;
     Cars cars;
 
     private final RandomNumberGenerater randomNumberGenerater = new RandomNumberGenerater();
 
-    public Race(Cars cars, int times) {
+    public Race(Cars cars, long times) {
         this.cars = cars;
         this.times = times;
         this.count = 0;
